requirements.txt
asgiref==3.4.1
Django==3.2.12
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.0.4
pytz==2021.3
sqlparse==0.4.2
xlwt==1.3.0
drf_yasg==1.20.0
pyAesCrypt==6.0.0
python-decouple==3.6
google-api-python-client==2.42.0
google-auth-oauthlib==0.5.1
